[FIX] Vllm guided decoding params (#3662)
* vllm sampling params fix

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* do not patch base_trainer

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* seperate vllm fixes

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Apply suggestion from @danielhanchen

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Revert "[pre-commit.ci] auto fixes from pre-commit.com hooks"

This reverts commit fe334c4dffda0c799295a69b3fc5f4c071d83176.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Revert "[pre-commit.ci] auto fixes from pre-commit.com hooks"

This reverts commit ae94be7d7e0a4780820adfbcde3549aec0ec49cf.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Revert "[pre-commit.ci] auto fixes from pre-commit.com hooks"

This reverts commit cde2dde170d35bce2b50211eeaf4c9788f33734c.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

+++ unsloth/kernels/layernorm.py
@@ -47,19 +47,19 @@ def layernorm_forward(
 
     # According to https://pytorch.org/torchtune/stable/_modules/torchtune/modules/layer_norm.html#Fp32LayerNorm, all modules
     # are in float32!
-    X_row = tl.load(X + col_offsets, mask = mask, other = 0).to(tl.float32)
-    W_row = tl.load(W + col_offsets, mask = mask, other = 0).to(tl.float32)
-    b_row = tl.load(b + col_offsets, mask = mask, other = 0).to(tl.float32)
+    X_row = tl.load(X + col_offsets, mask=mask, other=0).to(tl.float32)
+    W_row = tl.load(W + col_offsets, mask=mask, other=0).to(tl.float32)
+    b_row = tl.load(b + col_offsets, mask=mask, other=0).to(tl.float32)
 
-    mean_X = tl.sum(X_row, axis = 0) / n_cols
+    mean_X = tl.sum(X_row, axis=0) / n_cols
     # (X[0] - mean) == -mean so we need to mask it out
     XX = tl.where(mask, X_row - mean_X, 0)
-    row_var = tl.sum(XX * XX, axis = 0) / n_cols
+    row_var = tl.sum(XX * XX, axis=0) / n_cols
     inv_var = tl.math.rsqrt(row_var + eps)
     tl.store(r, inv_var)
     tl.store(mu, mean_X)
     output = (XX * inv_var) * W_row + b_row
-    tl.store(Y + col_offsets, output, mask = mask)
+    tl.store(Y + col_offsets, output, mask=mask)
 
 
 @triton.jit
@@ -88,10 +88,10 @@ def layernorm_backward(
 
     # According to https://pytorch.org/torchtune/stable/_modules/torchtune/modules/layer_norm.html#Fp32LayerNorm, all modules
     # are in float32!
-    dY_row = tl.load(dY + col_offsets, mask = mask, other = 0).to(tl.float32)
-    X_row = tl.load(X + col_offsets, mask = mask, other = 0).to(tl.float32)
-    W_row = tl.load(W + col_offsets, mask = mask, other = 0).to(tl.float32)
-    b_row = tl.load(b + col_offsets, mask = mask, other = 0).to(tl.float32)
+    dY_row = tl.load(dY + col_offsets, mask=mask, other=0).to(tl.float32)
+    X_row = tl.load(X + col_offsets, mask=mask, other=0).to(tl.float32)
+    W_row = tl.load(W + col_offsets, mask=mask, other=0).to(tl.float32)
+    b_row = tl.load(b + col_offsets, mask=mask, other=0).to(tl.float32)
 
     inv_var = tl.load(r).to(tl.float32)
     mean = tl.load(mu).to(tl.float32)
@@ -99,11 +99,11 @@ def layernorm_backward(
     dY_W = dY_row * W_row
     dX_row = (
         dY_W
-        - tl.sum(dY_W, axis = 0) / n_cols
-        - normed * tl.sum(dY_W * normed, axis = 0) / n_cols
+        - tl.sum(dY_W, axis=0) / n_cols
+        - normed * tl.sum(dY_W * normed, axis=0) / n_cols
     )
     dX_row = dX_row * inv_var
-    tl.store(dY + col_offsets, dX_row, mask = mask)
+    tl.store(dY + col_offsets, dX_row, mask=mask)
 
 
 class Fast_Layernorm(torch.autograd.Function):
@@ -115,9 +115,9 @@ class Fast_Layernorm(torch.autograd.Function):
         n_rows, n_cols = X.shape
         BLOCK_SIZE, num_warps = calculate_settings(n_cols)
         device = X.device
-        Y = torch.empty((n_rows, n_cols), dtype = X.dtype, device = device)
-        r = torch.empty(n_rows, dtype = torch.float32, device = device)
-        mu = torch.empty(n_rows, dtype = torch.float32, device = device)
+        Y = torch.empty((n_rows, n_cols), dtype=X.dtype, device=device)
+        r = torch.empty(n_rows, dtype=torch.float32, device=device)
+        mu = torch.empty(n_rows, dtype=torch.float32, device=device)
 
         with torch_gpu_device(device):
             layernorm_forward[(n_rows,)](
@@ -131,8 +131,8 @@ class Fast_Layernorm(torch.autograd.Function):
                 mu,
                 n_cols,
                 eps,
-                BLOCK_SIZE = BLOCK_SIZE,
-                num_warps = num_warps,
+                BLOCK_SIZE=BLOCK_SIZE,
+                num_warps=num_warps,
             )
         ctx.eps = eps
         ctx.BLOCK_SIZE = BLOCK_SIZE
@@ -160,8 +160,8 @@ class Fast_Layernorm(torch.autograd.Function):
                 mu,
                 n_cols,
                 ctx.eps,
-                BLOCK_SIZE = ctx.BLOCK_SIZE,
-                num_warps = ctx.num_warps,
+                BLOCK_SIZE=ctx.BLOCK_SIZE,
+                num_warps=ctx.num_warps,
             )
         dX = dY.view(*shape)
         return dX, None, None, None, None
@@ -181,26 +181,26 @@ def fast_layernorm(layernorm, X):
 
 
 def test_layernorm(
-    dim = 1024,
-    eps = 1e-5,
-    dtype = torch.float16,
-    bsz = 21,
-    random_state = 3407,
-    seqlen = 3341,
+    dim=1024,
+    eps=1e-5,
+    dtype=torch.float16,
+    bsz=21,
+    random_state=3407,
+    seqlen=3341,
 ):
     from torch.nn import LayerNorm
 
-    layernorm = LayerNorm((dim,), eps = eps, device = "cuda", dtype = dtype)
+    layernorm = LayerNorm((dim,), eps=eps, device="cuda", dtype=dtype)
     torch.cuda.manual_seed(random_state)
     torch.manual_seed(random_state)
     torch.nn.init.uniform_(layernorm.weight)
     torch.nn.init.uniform_(layernorm.bias)
-    X = torch.randn((bsz, seqlen, dim), dtype = dtype, device = "cuda")
+    X = torch.randn((bsz, seqlen, dim), dtype=dtype, device="cuda")
     XX = X.clone()
     X.requires_grad_(True)
     XX.requires_grad_(True)
     Y = layernorm(X)
-    YY = torch.randn((bsz, seqlen, dim), dtype = dtype, device = "cuda", requires_grad = True)
+    YY = torch.randn((bsz, seqlen, dim), dtype=dtype, device="cuda", requires_grad=True)
     Y.backward(YY)
     correct_grad = X.grad.clone()
     # from unsloth.kernels import fast_layernorm
@@ -212,14 +212,14 @@ def test_layernorm(
 def testing_suite_layernorm():
     for dim in [512, 1024, 2048]:
         for dtype in [torch.float16, torch.bfloat16]:
-            with torch.autocast(device_type = "cuda", dtype = dtype):
+            with torch.autocast(device_type="cuda", dtype=dtype):
                 for seqlen in [3341, 2048, 349]:
                     for random_state in [3407, 42]:
                         test_layernorm(
-                            dim = dim,
-                            eps = 1e-5,
-                            dtype = dtype,
-                            bsz = 21,
-                            random_state = random_state,
-                            seqlen = seqlen,
+                            dim=dim,
+                            eps=1e-5,
+                            dtype=dtype,
+                            bsz=21,
+                            random_state=random_state,
+                            seqlen=seqlen,
                         )
